% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODBINOM.R
\name{dprodbinom}
\alias{dprodbinom}
\alias{pprodbinom}
\alias{qprodbinom}
\alias{rprodbinom}
\title{Product of Independent Binomail Distributions}
\usage{
dprodbinom(x, size, prob, log = FALSE)

pprodbinom(q, size, prob, lower.tail = TRUE, log.p = FALSE)

qprodbinom(p, size, prob, lower.tail = TRUE, log.p = FALSE)

rprodbinom(n, size, prob)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{size}{vector of trials (zero or more).}

\item{prob}{vector of probability of success on each trial.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{lower.tails}{logical; if \code{TRUE} (default), probabilities are P[X <= x], otherwise, P[X > x].}
}
\description{
Density, distribution, quantile, and random generation functions for products of independent binomial distributions.
}
