% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRODBINOM-package.R
\docType{package}
\name{PRODBINOM}
\alias{PRODBINOM}
\title{Product of Independent Binomail Distributions}
\description{
Density, distribution, quantile, and random generation functions for products of independent binomial distributions.
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
